+++ aidial_adapter_openai/dial_api/resource.py
